requirements.txt
flask>=0.12.2
flask_graphql>=1.4
graphene>=2.0
psycopg2>=2.7.3
promise>=2.1
redis>=2.10
google-auth>=1.3
requests>=2.18
expiringdict>=1.1.4
